Terminate gatsby usage
Delete all gatsby files, and continue development with only
'vanilla' react. All react files migrated to the parent directory.

Note that this was the original inteded direction of the project,
with gatsby intended as an experimental sideline for static
elements of the service, such as periodic blog posts. However,
this approach seems to entail more work than is practical at
the current time.